Move initializing analytics into a separate function.
